refactor(layout): 将内建方法的内容单独抽出来。进行抽象
